Se modularizo la app y se acomodo para luego dar estilos
