+++ app-clima-matias-torres/pages/index.js
@@ -2,8 +2,8 @@
 import Head from "next/head";
 import { Inter } from "next/font/google";
 import styles from "@/styles/Home.module.css";
-import InputCiudad from "@/components/InputCiudad";
-import { useState } from "react";
+
+import Clima from "@/components/Clima";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -19,8 +19,8 @@ export default function Home() {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <main className={`${styles.main} ${inter.className}`}>
-        <InputCiudad />
+      <main className={`${inter.className}`}>
+        <Clima />
       </main>
     </>
   );
